refactor: atualiza import de cores gráfico donut
